Prepare backend for remote

See merge request cs309/fall2023/ip_106!9

+++ Backend/geoquizbackend/src/main/java/com/example/geoquizbackend/Quiz/QuizController.java
@@ -21,7 +21,7 @@ public class QuizController {
     }
 
     @GetMapping(path = "quizzes/{id}")
-    Quiz getQuizById(@PathVariable int id) {
+    Quiz getQuizById(@PathVariable long id) {
         return quizRepository.findById(id);
     }
